refactor(inspection-service): extract inspection-photo sub-service
Move R2 photo upload (EXIF strip), media-center aggregation, loose-pool
upload/attach/delete, item-photo reorder/detach/revert/move, the field-version
-aware patchItem, and the isInspectionPhotoKey cover-validation predicate into
InspectionPhotoService. Ownership checks call back through the facade
(getInspection). The facade delegates; moved code is byte-identical. The facade
no longer stores r2/images as fields (passed straight to sub-services).
